Extract class for reading bcrs
